==> script/S02.find_peak/07.combine_greenwave_and_birdwave_peak.ipynb <==
import pandas as pd
import numpy as np


data1 = pd.read_csv('../../data/D02.wave_peak/all_birdwave_peak.csv')
data2 = pd.read_csv('../../data/D02.wave_peak/all_greenwave_peak.csv', index_col=0)
data2['niche_or_level'] = 'peak_IRG'


data = pd.concat([data1, data2], axis=0, ignore_index=True).reset_index(drop=True)

data.to_csv('../../data/D02.wave_peak/all_birdwave_greenwave_peak.csv', index=False)
